// stencil/promo/src/components/cnt-flexy-view-islam-promo-footer/res/view/cnt-flexy-view-islam-promo-footer/interface/common.interface.ts
/**
 * Элементы компонента Cnt-flexy-view-islam-promo-footer
 */
export interface SCntFlexyViewIslamPromoFooter{
  id: any;
  maleIcon: any;
  contact: string;
  phoneIcon: any;
  phone: string;
  letterIcon: any;
  letter: string;
  mapMarker: any;
  location: string;
  vkIcon: any;
  twitterIcon: any;
  facebookIcon: any;
  instagramIcon: any;
  youtubeIcon: any;
  link: string;
  homeNav: string;
  aboutUsNav: string;
  ourTeamNav: string;
  contactsNav: string;
  email: string;
  button: string;
}